refactor: move CheckError to top-level check module
Move CheckError and its From impls out of the csaf-gated verification
module into a new crate::check module that is always available. This
fixes cargo check-all-features which builds without the csaf feature.
Re-export from verification::check for backward compatibility.
